# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cottage Village API Documentation",
        "description": "",
        "contact": {
            "email": "afonico@info.ru"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/news": {
            "post": {
                "tags": [
                    "News"
                ],
                "summary": "Get news posts",
                "description": "",
                "operationId": "index",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "limit": {
                                        "description": "News limit. Default value 8",
                                        "type": "integer"
                                    },
                                    "offset": {
                                        "description": "News offset. Default value 0",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return token object { 'access_token', 'token_type', 'expires_in' }"
                    }
                }
            }
        },
        "/api/news/single": {
            "post": {
                "tags": [
                    "News"
                ],
                "summary": "Get single news post",
                "description": "",
                "operationId": "single",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "news_id"
                                ],
                                "properties": {
                                    "news_id": {
                                        "description": "News id",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return token object { 'access_token', 'token_type', 'expires_in' }"
                    }
                }
            }
        },
        "/api/news/create": {
            "post": {
                "tags": [
                    "News"
                ],
                "summary": "Create news post",
                "description": "",
                "operationId": "create",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "logo",
                                    "subtitle",
                                    "content"
                                ],
                                "properties": {
                                    "logo": {
                                        "description": "News logo image",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "subtitle": {
                                        "description": "News subtitle",
                                        "type": "string"
                                    },
                                    "content": {
                                        "description": "News content",
                                        "type": "string"
                                    },
                                    "files": {
                                        "description": "News attachments",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return token object { 'access_token', 'token_type', 'expires_in' }"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/news/update": {
            "post": {
                "tags": [
                    "News"
                ],
                "summary": "Update news post",
                "description": "files paramerer only creates new files. If files not changed, u should not post this parameter",
                "operationId": "update",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "description": "News id",
                                        "type": "integer"
                                    },
                                    "logo": {
                                        "description": "News logo image",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "subtitle": {
                                        "description": "News subtitle",
                                        "type": "string"
                                    },
                                    "content": {
                                        "description": "News content",
                                        "type": "string"
                                    },
                                    "deleted_files[]": {
                                        "description": "Attachments ids array to delete files with posts",
                                        "type": "integer"
                                    },
                                    "files": {
                                        "description": "News attachments",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/news/delete": {
            "post": {
                "tags": [
                    "News"
                ],
                "summary": "Delete news post with attachments",
                "description": "",
                "operationId": "delete",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "news_id"
                                ],
                                "properties": {
                                    "news_id": {
                                        "description": "News id",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/news/delete/file": {
            "post": {
                "tags": [
                    "News"
                ],
                "summary": "Delete news post attachment",
                "description": "",
                "operationId": "deleteFile",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "attachment_id"
                                ],
                                "properties": {
                                    "attachment_id": {
                                        "description": "Attachment id",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format: token_type access_token",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}